@@ -30,23 +30,26 @@
     "rulesets/*"
   ],
   "engines": {
-    "node": ">=8.3.0"
+    "node": ">=10.0"
   },
   "scripts": {
     "build.binary": "pkg . --output ./binaries/spectral",
+    "build.clean": "rimraf ./coverage && rimraf ./dist && rimraf ./rulesets && rimraf ./__karma__/__fixtures__",
     "build.oas-functions": "rollup -c",
     "build": "tsc -p ./tsconfig.build.json",
     "cli": "node -r ts-node/register -r tsconfig-paths/register src/cli/index.ts",
     "cli:debug": "node -r ts-node/register -r tsconfig-paths/register --inspect-brk src/cli/index.ts",
-    "compile-rulesets": "node ./scripts/compile-rulesets.js",
+    "generate-assets": "node ./scripts/generate-assets.js",
     "inline-version": "./scripts/inline-version.js",
     "lint.fix": "yarn lint --fix",
-    "lint": "tslint 'src/**/*.ts'",
+    "lint": "tsc --noEmit && tslint 'src/**/*.ts'",
+    "copy.html-templates": "copyfiles -u 1 \"./src/formatters/html/*.html\" \"./dist/\"",
     "postbuild.oas-functions": "copyfiles -u 1 \"dist/rulesets/oas*/functions/*.js\" ./",
-    "postbuild": "yarn build.oas-functions && yarn compile-rulesets",
-    "prebuild": "copyfiles -u 1 \"src/rulesets/oas*/**/*.json\" dist && copyfiles -u 1 \"src/rulesets/oas*/**/*.json\" ./",
+    "postbuild": "yarn build.oas-functions && yarn generate-assets",
+    "prebuild": "yarn build.clean && copyfiles -u 1 \"src/rulesets/oas*/**/*.json\" dist && copyfiles -u 1 \"src/rulesets/oas*/**/*.json\" ./ && yarn copy.html-templates",
     "prebuild.binary": "yarn build",
-    "pretest.karma": "node ./scripts/generate-karma-fixtures.js",
+    "pretest.karma": "node ./scripts/generate-karma-fixtures.js && yarn pretest",
+    "pretest": "node ./scripts/generate-assets.js",
     "schema.update": "yarn typescript-json-schema --id \"http://stoplight.io/schemas/rule.schema.json\" --required tsconfig.json IRule --out ./src/meta/rule.schema.json",
     "test.harness": "jest -c ./jest.harness.config.js",
     "test.karma": "karma start",
@@ -56,70 +59,75 @@
     "test": "jest --silent"
   },
   "dependencies": {
-    "@stoplight/json": "^3.1.1",
-    "@stoplight/json-ref-resolver": "^2.2.0",
-    "@stoplight/path": "^1.2.0",
-    "@stoplight/types": "^11.0.0",
-    "@stoplight/yaml": "^3.1.0",
+    "@stoplight/json": "^3.4.0",
+    "@stoplight/json-ref-readers": "^1.1.1",
+    "@stoplight/json-ref-resolver": "^3.0.6",
+    "@stoplight/path": "^1.3.0",
+    "@stoplight/types": "^11.2.0",
+    "@stoplight/yaml": "^3.5.0",
     "abort-controller": "^3.0.0",
-    "ajv": "^6.7",
-    "ajv-oai": "^1.1.1",
+    "ajv": "^6.10",
+    "ajv-oai": "^1.1.5",
     "better-ajv-errors": "^0.6.7",
-    "chalk": "^2.4.2",
-    "deprecated-decorator": "^0.1.6",
-    "fast-glob": "^3.0.4",
-    "jsonpath-plus": "~1.0",
+    "blueimp-md5": "^2.12.0",
+    "chalk": "^3.0.0",
+    "eol": "^0.9.1",
+    "fast-glob": "^3.1.0",
+    "jsonpath-plus": "~2.0",
     "lodash": ">=4.17.5",
-    "nanoid": "^2.0.3",
+    "nanoid": "^2.1.6",
     "node-fetch": "^2.6",
-    "proxy-agent": "^3.1.0",
-    "strip-ansi": "^5.2",
+    "proxy-agent": "^3.1.1",
+    "strip-ansi": "^6.0",
     "text-table": "^0.2",
     "tslib": "^1.10.0",
-    "typescript-json-schema": "~0.40",
-    "yargs": "^14.0.0"
+    "yargs": "^15.0.2"
   },
   "devDependencies": {
-    "@commitlint/config-conventional": "^8.1.0",
+    "@commitlint/config-conventional": "^8.2.0",
     "@types/chalk": "^2.2.0",
     "@types/fetch-mock": "^7.3.1",
-    "@types/jest": "^24.0.16",
-    "@types/jest-when": "^2.4.0",
-    "@types/lodash": "^4.14.136",
-    "@types/nanoid": "^2.0.0",
+    "@types/jest": "^24.0.22",
+    "@types/jest-when": "^2.7.0",
+    "@types/lodash": "^4.14.146",
+    "@types/nanoid": "^2.1.0",
     "@types/nock": "^11.1.0",
-    "@types/node": "~12.7",
-    "@types/node-fetch": "^2.5.0",
-    "@types/text-table": "^0.2.0",
-    "@types/xml2js": "^0.4.4",
-    "@types/yaml": "^1.0.2",
+    "@types/node": "~12.12",
+    "@types/node-fetch": "^2.5.3",
+    "@types/text-table": "^0.2.1",
+    "@types/tmp": "^0.1.0",
+    "@types/xml2js": "^0.4.5",
     "copyfiles": "^2.1.1",
-    "fetch-mock": "^7.3.9",
+    "dependency-graph": "^0.8.0",
+    "fetch-mock": "^7.7.3",
     "glob-fs": "^0.1.7",
-    "husky": "^3.0.0",
+    "husky": "^3.0.9",
     "jest": "~24.9",
-    "jest-mock": "^24.8.0",
+    "jest-mock": "^24.9.0",
     "jest-when": "~2.7",
-    "karma": "^4.2.0",
-    "karma-chrome-launcher": "^3.0.0",
+    "karma": "^4.4.1",
+    "karma-chrome-launcher": "^3.1.0",
     "karma-jasmine": "^2.0.1",
     "karma-typescript": "^4.1.1",
-    "lint-staged": "^9.0.2",
-    "nock": "~11.4.0",
+    "lint-staged": "^9.4.2",
+    "nock": "~11.7.0",
+    "node-html-parser": "^1.1.16",
     "pkg": "^4.4.0",
-    "recast": "^0.18.1",
-    "rollup": "^1.19.4",
-    "rollup-plugin-commonjs": "^10.0.2",
+    "recast": "^0.18.5",
+    "rimraf": "^3.0.0",
+    "rollup": "^1.27.14",
+    "rollup-plugin-commonjs": "^10.1.0",
     "rollup-plugin-node-resolve": "^5.2.0",
-    "rollup-plugin-terser": "^5.1.1",
-    "rollup-plugin-typescript2": "^0.24.0",
-    "ts-jest": "^24.0.2",
-    "ts-node": "^8.1.0",
-    "tsconfig-paths": "^3.8.0",
+    "rollup-plugin-terser": "^5.1.2",
+    "rollup-plugin-typescript2": "^0.25.2",
+    "ts-jest": "^24.1.0",
+    "ts-node": "^8.4.1",
+    "tsconfig-paths": "^3.9.0",
     "tslint": "~5.20",
-    "tslint-config-stoplight": "~1.3",
-    "typescript": "^3.5.3",
-    "xml2js": "^0.4.21"
+    "tslint-config-stoplight": "~1.4",
+    "typescript": "^3.7.4",
+    "typescript-json-schema": "~0.41.0",
+    "xml2js": "^0.4.23"
   },
   "lint-staged": {
     "*.{ts,tsx}": [
