@@ -1,14 +1,12 @@
-import { JsonPath } from '@stoplight/types';
-import { get, has } from 'lodash';
-
-const { JSONPath } = require('jsonpath-plus');
-
-import { decodePointerFragment, pathToPointer } from '@stoplight/json';
-import { Resolved } from './resolved';
-import { message } from './rulesets/message';
+import { decodePointerFragment } from '@stoplight/json';
+import { get } from 'lodash';
+import { getDefaultRange, Resolved } from './resolved';
+import { IMessageVars, message } from './rulesets/message';
 import { getDiagnosticSeverity } from './rulesets/severity';
 import { IFunction, IGivenNode, IRuleResult, IRunRule, IThen } from './types';
-import { isObject } from './utils';
+import { getClosestJsonPath, printPath, PrintStyle } from './utils';
+
+const { JSONPath } = require('jsonpath-plus');
 
 // TODO(SO-23): unit test but mock whatShouldBeLinted
 export const lintNode = (
@@ -19,14 +17,7 @@ export const lintNode = (
   resolved: Resolved,
 ): IRuleResult[] => {
   const givenPath = node.path[0] === '$' ? node.path.slice(1) : node.path;
-  const conditioning = whatShouldBeLinted(givenPath, node.value, rule);
-
-  // If the 'when' condition is not satisfied, simply don't run the linter
-  if (!conditioning.lint) {
-    return [];
-  }
-
-  const targetValue = conditioning.value;
+  const targetValue = node.value;
 
   const targets: any[] = [];
   if (then && then.field) {
@@ -75,7 +66,7 @@ export const lintNode = (
     });
   }
 
-  let results: IRuleResult[] = [];
+  const results: IRuleResult[] = [];
 
   for (const target of targets) {
     const targetPath = givenPath.concat(target.path);
@@ -91,44 +82,42 @@ export const lintNode = (
         {
           original: node.value,
           given: node.value,
+          resolved,
         },
       ) || [];
 
-    results = results.concat(
-      targetResults.map<IRuleResult>(result => {
+    results.push(
+      ...targetResults.map<IRuleResult>(result => {
         const escapedJsonPath = (result.path || targetPath).map(segment => decodePointerFragment(String(segment)));
-        const path = getClosestJsonPath(
-          rule.resolved === false ? resolved.unresolved : resolved.resolved,
-          escapedJsonPath,
-        );
-        // todo: https://github.com/stoplightio/spectral/issues/608
-        const location = resolved.getLocationForJsonPath(path, true);
+        const parsed = resolved.getParsedForJsonPath(escapedJsonPath);
+        const path = parsed?.path || getClosestJsonPath(resolved.resolved, escapedJsonPath);
+        const doc = parsed?.doc || resolved.parsed;
+        const range = doc.getLocationForJsonPath(doc.parsed, path, true)?.range || getDefaultRange();
+        const value = path.length === 0 ? parsed?.doc.parsed.data : get(parsed?.doc.parsed.data, path);
+
+        const vars: IMessageVars = {
+          property:
+            parsed?.missingPropertyPath && parsed.missingPropertyPath.length > path.length
+              ? printPath(parsed.missingPropertyPath.slice(path.length - 1), PrintStyle.Dot)
+              : path.length > 0
+              ? path[path.length - 1]
+              : '',
+          error: result.message,
+          path: printPath(path, PrintStyle.EscapedPointer),
+          description: rule.description,
+          value,
+        };
+
+        const resultMessage = message(result.message, vars);
+        vars.error = resultMessage;
 
         return {
           code: rule.name,
-
-          message:
-            rule.message === undefined
-              ? rule.description || result.message
-              : message(rule.message, {
-                  error: result.message,
-                  property: path.length > 0 ? path[path.length - 1] : '',
-                  path: pathToPointer(path),
-                  description: rule.description,
-                  get value() {
-                    // let's make it `value` lazy
-                    const value = resolved.getValueForJsonPath(path);
-                    if (isObject(value)) {
-                      return Array.isArray(value) ? 'Array[]' : 'Object{}';
-                    }
-
-                    return JSON.stringify(value);
-                  },
-                }),
+          message: (rule.message === void 0 ? rule.description ?? resultMessage : message(rule.message, vars)).trim(),
           path,
           severity: getDiagnosticSeverity(rule.severity),
-          source: location.uri,
-          range: location.range,
+          source: parsed?.doc.source,
+          range,
         };
       }),
     );
@@ -136,88 +125,3 @@ export const lintNode = (
 
   return results;
 };
-
-// TODO(SO-23): unit test idividually
-export const whatShouldBeLinted = (
-  path: JsonPath,
-  originalValue: any,
-  rule: IRunRule,
-): { lint: boolean; value: any } => {
-  const leaf = path[path.length - 1];
-
-  const when = rule.when;
-  if (!when) {
-    return {
-      lint: true,
-      value: originalValue,
-    };
-  }
-
-  const pattern = when.pattern;
-  const field = when.field;
-
-  // TODO: what if someone's field is called '@key'? should we use @@key?
-  const isKey = field === '@key';
-
-  if (!pattern) {
-    // isKey doesn't make sense without pattern
-    if (isKey) {
-      return {
-        lint: false,
-        value: originalValue,
-      };
-    }
-
-    return {
-      lint: has(originalValue, field),
-      value: originalValue,
-    };
-  }
-
-  if (isKey && pattern) {
-    return keyAndOptionalPattern(leaf, pattern, originalValue);
-  }
-
-  const fieldValue = String(get(originalValue, when.field));
-
-  return {
-    lint: fieldValue.match(pattern) !== null,
-    value: originalValue,
-  };
-};
-
-function keyAndOptionalPattern(key: string | number, pattern: string, value: any) {
-  /** arrays, look at the keys on the array object. note, this number check on id prop is not foolproof... */
-  if (typeof key === 'number' && typeof value === 'object') {
-    for (const k of Object.keys(value)) {
-      if (String(k).match(pattern)) {
-        return {
-          lint: true,
-          value,
-        };
-      }
-    }
-  } else if (String(key).match(pattern)) {
-    // objects
-    return {
-      lint: true,
-      value,
-    };
-  }
-
-  return {
-    lint: false,
-    value,
-  };
-}
-
-// todo: revisit -> https://github.com/stoplightio/spectral/issues/608
-function getClosestJsonPath(data: unknown, path: JsonPath) {
-  if (data === null || typeof data !== 'object') return [];
-
-  while (path.length > 0 && !has(data, path)) {
-    path.pop();
-  }
-
-  return path;
-}
