@@ -5,9 +5,10 @@ import * as nock from 'nock';
 import * as path from 'path';
 import { isOpenApiv2 } from '../formats';
 import { httpAndFileResolver } from '../resolvers/http-and-file';
-import { Spectral } from '../spectral';
+import { IRunRule, Spectral } from '../spectral';
 
 const oasRuleset = require('../rulesets/oas/index.json');
+const oasRulesetRules: Dictionary<IRunRule, string> = oasRuleset.rules;
 const customOASRuleset = require('./__fixtures__/custom-oas-ruleset.json');
 
 describe('Spectral', () => {
@@ -22,7 +23,7 @@ describe('Spectral', () => {
 
       expect(s.rules).toEqual(
         expect.objectContaining({
-          ...[...Object.entries(oasRuleset.rules)].reduce<Dictionary<unknown>>((oasRules, [name, rule]) => {
+          ...[...Object.entries(oasRulesetRules)].reduce<Dictionary<IRunRule, string>>((oasRules, [name, rule]) => {
             oasRules[name] = {
               name,
               ...rule,
@@ -71,13 +72,32 @@ describe('Spectral', () => {
         'info-matches-stoplight': {
           ...ruleset.rules['info-matches-stoplight'],
           name: 'info-matches-stoplight',
-          severity: -1,
+          severity: DiagnosticSeverity.Warning,
         },
       });
     });
   });
 
-  test('reports issues for correct files with correct ranges and paths', async () => {
+  test('should support combining built-in ruleset with a custom one', async () => {
+    const s = new Spectral();
+    await s.loadRuleset(['spectral:oas', path.join(__dirname, './__fixtures__/bare-oas-ruleset.json')]);
+
+    expect(Object.values(s.rules)).toEqual(
+      expect.arrayContaining([
+        expect.objectContaining({
+          name: 'oas2-schema',
+        }),
+        expect.objectContaining({
+          name: 'oas3-schema',
+        }),
+        expect.objectContaining({
+          name: 'info-matches-stoplight',
+        }),
+      ]),
+    );
+  });
+
+  test('should report issues for correct files with correct ranges and paths', async () => {
     const documentUri = path.join(__dirname, './__fixtures__/document-with-external-refs.oas2.json');
     const spectral = new Spectral({ resolver: httpAndFileResolver });
     await spectral.loadRuleset('spectral:oas');
@@ -85,6 +105,7 @@ describe('Spectral', () => {
     const parsed = {
       parsed: parseWithPointers(fs.readFileSync(documentUri, 'utf8')),
       getLocationForJsonPath,
+      source: documentUri,
     };
 
     const results = await spectral.run(parsed, {
@@ -108,11 +129,11 @@ describe('Spectral', () => {
               line: 16,
             },
           },
-          source: undefined,
+          source: documentUri,
         }),
         expect.objectContaining({
           code: 'oas2-schema',
-          path: ['paths', '/todos/{todoId}', 'get', 'responses', '200', 'schema'],
+          path: [],
           range: {
             end: {
               character: 1,
@@ -138,7 +159,7 @@ describe('Spectral', () => {
               line: 10,
             },
           },
-          source: undefined,
+          source: documentUri,
         }),
         expect.objectContaining({
           code: 'info-contact',
@@ -153,7 +174,7 @@ describe('Spectral', () => {
               line: 2,
             },
           },
-          source: undefined,
+          source: documentUri,
         }),
         expect.objectContaining({
           code: 'operation-description',
@@ -168,9 +189,78 @@ describe('Spectral', () => {
               line: 11,
             },
           },
-          source: undefined,
+          source: documentUri,
         }),
       ]),
     );
+  });
+
+  test('should recognize the source of remote $refs, and de-dupe results by fingerprint', async () => {
+    const s = new Spectral({ resolver: httpAndFileResolver });
+    const documentUri = path.join(__dirname, './__fixtures__/gh-658/URIError.yaml');
+
+    s.setRules({
+      'schema-strings-maxLength': {
+        severity: DiagnosticSeverity.Warning,
+        recommended: true,
+        message: "String typed properties MUST be further described using 'maxLength'. Error: {{error}}",
+        given: "$..[?(@.type === 'string')]",
+        then: {
+          field: 'maxLength',
+          function: 'truthy',
+        },
+      },
+    });
+
+    const results = await s.run(fs.readFileSync(documentUri, 'utf8'), { resolve: { documentUri } });
+
+    expect(results.length).toEqual(3);
+
+    return expect(results).toEqual([
+      expect.objectContaining({
+        path: ['components', 'schemas', 'Error', 'properties', 'status_code'],
+        source: expect.stringContaining('/src/__tests__/__fixtures__/gh-658/lib.yaml'),
+        range: {
+          end: {
+            character: 22,
+            line: 21,
+          },
+          start: {
+            character: 20,
+            line: 20,
+          },
+        },
+      }),
+
+      expect.objectContaining({
+        path: ['paths', '/test', 'get', 'responses', '200', 'content', 'application/json', 'schema'],
+        source: expect.stringContaining('/src/__tests__/__fixtures__/gh-658/URIError.yaml'),
+        range: {
+          end: {
+            character: 28,
+            line: 23,
+          },
+          start: {
+            character: 21,
+            line: 22,
+          },
+        },
+      }),
+
+      expect.objectContaining({
+        path: ['components', 'schemas', 'Foo'],
+        source: expect.stringContaining('/src/__tests__/__fixtures__/gh-658/URIError.yaml'),
+        range: {
+          end: {
+            character: 18,
+            line: 43,
+          },
+          start: {
+            character: 8,
+            line: 42,
+          },
+        },
+      }),
+    ]);
   });
 });
