@@ -1,3 +1,4 @@
+import { DiagnosticSeverity } from '@stoplight/types/dist';
 import { FetchMockSandbox } from 'fetch-mock';
 import { Spectral } from '../spectral';
 
@@ -44,7 +45,7 @@ describe('Spectral', () => {
         'info-matches-stoplight': {
           ...ruleset.rules['info-matches-stoplight'],
           name: 'info-matches-stoplight',
-          severity: -1,
+          severity: DiagnosticSeverity.Warning,
         },
       });
     });
