@@ -20,6 +20,7 @@ function run(command: string) {
 
 describe('lint', () => {
   let errorSpy: jest.SpyInstance;
+  const { isTTY } = process.stdin;
 
   const results: IRuleResult[] = [
     {
@@ -55,11 +56,29 @@ describe('lint', () => {
 
   afterEach(() => {
     errorSpy.mockRestore();
+    process.stdin.isTTY = isTTY;
   });
 
-  it('shows help when no document argument is passed', async () => {
+  it('shows help when no document and no STDIN are present', async () => {
+    process.stdin.isTTY = true;
     const output = await run('lint');
     expect(output).toContain('documents  Location of JSON/YAML documents');
+  });
+
+  describe('when STDIN is present', () => {
+    it('does not show help when documents are missing', async () => {
+      const output = await run('lint');
+      expect(output).not.toContain('documents  Location of JSON/YAML documents');
+    });
+
+    it('calls with lint with STDIN file descriptor', async () => {
+      await run('lint');
+      expect(lint).toBeCalledWith([0], {
+        encoding: 'utf8',
+        format: 'stylish',
+        ignoreUnknownFormat: false,
+      });
+    });
   });
 
   it('shows help when invalid arguments are passed', async () => {
@@ -73,6 +92,7 @@ describe('lint', () => {
     expect(lint).toBeCalledWith([doc], {
       encoding: 'utf8',
       format: 'stylish',
+      ignoreUnknownFormat: false,
     });
   });
 
@@ -82,6 +102,7 @@ describe('lint', () => {
     expect(lint).toBeCalledWith([doc], {
       encoding: 'utf16',
       format: 'stylish',
+      ignoreUnknownFormat: false,
     });
   });
 
@@ -91,6 +112,7 @@ describe('lint', () => {
     expect(lint).toBeCalledWith([doc], {
       encoding: 'utf16',
       format: 'json',
+      ignoreUnknownFormat: false,
     });
   });
 
@@ -147,6 +169,16 @@ describe('lint', () => {
       skipRule: ['foo', 'bar'],
       encoding: 'utf8',
       format: 'stylish',
+      ignoreUnknownFormat: false,
+    });
+  });
+
+  it('passes ignore-unknown-format to lint', async () => {
+    await run('lint --ignore-unknown-format ./__fixtures__/empty-oas2-document.json');
+    expect(lint).toHaveBeenCalledWith([expect.any(String)], {
+      encoding: 'utf8',
+      format: 'stylish',
+      ignoreUnknownFormat: true,
     });
   });
 
@@ -156,7 +188,7 @@ describe('lint', () => {
     );
   });
 
-  it('errors upon exception', async () => {
+  it('prints error message upon exception', async () => {
     const error = new Error('Failure');
     (lint as jest.Mock).mockReset();
     (lint as jest.Mock).mockReturnValueOnce({
@@ -171,6 +203,6 @@ describe('lint', () => {
     });
 
     await run(`lint -o foo.json ./__fixtures__/empty-oas2-document.json`);
-    expect(errorSpy).toBeCalledWith(error);
+    expect(errorSpy).toBeCalledWith('Failure');
   });
 });
