@@ -1,11 +1,17 @@
 import { normalize } from '@stoplight/path';
 import * as fg from 'fast-glob';
 
-export async function listFiles(pattens: string[]): Promise<string[]> {
-  const { files, urls } = pattens.reduce<{ files: string[]; urls: string[] }>(
+export async function listFiles(pattens: Array<number | string>): Promise<Array<number | string>> {
+  const { files, fileDescriptors, urls } = pattens.reduce<{
+    files: string[];
+    urls: string[];
+    fileDescriptors: number[];
+  }>(
     (group, pattern) => {
-      if (!/^https?:\/\//.test(pattern)) {
-        group.files.push(pattern);
+      if (typeof pattern === 'number') {
+        group.fileDescriptors.push(pattern);
+      } else if (!/^https?:\/\//.test(pattern)) {
+        group.files.push(pattern.replace(/\\/g, '/'));
       } else {
         group.urls.push(pattern);
       }
@@ -15,8 +21,9 @@ export async function listFiles(pattens: string[]): Promise<string[]> {
     {
       files: [],
       urls: [],
+      fileDescriptors: [],
     },
   );
 
-  return [...urls, ...(await fg(files, { dot: true, absolute: true })).map(normalize)]; // let's normalize OS paths produced by fast-glob to have consistent paths across all platforms
+  return [...urls, ...fileDescriptors, ...(await fg(files, { dot: true, absolute: true })).map(normalize)]; // let's normalize OS paths produced by fast-glob to have consistent paths across all platforms
 }
