@@ -1,7 +1,5 @@
-import { JSONSchema4, JSONSchema6, JSONSchema7 } from 'json-schema';
+import { JSONSchema } from '../types';
 import { isObject } from '../utils/isObject';
-
-type JSONSchema = JSONSchema4 | JSONSchema6 | JSONSchema7;
 
 const KNOWN_JSON_SCHEMA_TYPES = ['array', 'boolean', 'integer', 'null', 'number', 'object', 'string'];
 const KNOWN_JSON_SCHEMA_COMBINERS = ['allOf', 'oneOf', 'anyOf'];
