@@ -1,8 +1,9 @@
-import chalk from 'chalk';
+import * as chalk from 'chalk';
+import { sortResults } from '../../utils';
 import { stylish } from '../stylish';
 
-const oas3SchemaErrors = require('./__fixtures__/oas3-schema-errors.json');
-const mixedErrors = require('./__fixtures__/mixed-errors.json');
+const oas3SchemaErrors = sortResults(require('./__fixtures__/oas3-schema-errors.json'));
+const mixedErrors = sortResults(require('./__fixtures__/mixed-errors.json'));
 
 describe('Stylish formatter', () => {
   test('should prefer message for oas-schema errors', () => {
