@@ -1,6 +1,7 @@
 import { isURL } from '@stoplight/path';
 import AbortController from 'abort-controller';
 import * as fs from 'fs';
+import { STATIC_ASSETS } from '../assets';
 import request from '../request';
 
 export interface IReadOptions {
@@ -8,8 +9,31 @@ export interface IReadOptions {
   timeout?: number;
 }
 
-export async function readFile(name: string, opts: IReadOptions): Promise<string> {
-  if (isURL(name)) {
+export async function readFile(name: string | number, opts: IReadOptions): Promise<string> {
+  if (typeof name === 'number') {
+    let result = '';
+
+    const stream = fs.createReadStream('', { fd: name });
+    stream.setEncoding(opts.encoding);
+
+    stream.on('readable', () => {
+      let chunk: string | null;
+
+      // tslint:disable-next-line:no-conditional-assignment
+      while ((chunk = stream.read()) !== null) {
+        result += chunk;
+      }
+    });
+
+    return new Promise<string>((resolve, reject) => {
+      stream.on('error', reject);
+      stream.on('end', () => {
+        resolve(result);
+      });
+    });
+  } else if (name in STATIC_ASSETS) {
+    return STATIC_ASSETS[name];
+  } else if (isURL(name)) {
     let response;
     let timeout: NodeJS.Timeout | number | null = null;
     try {
@@ -53,7 +77,7 @@ export async function readFile(name: string, opts: IReadOptions): Promise<string
   }
 }
 
-export async function readParsable(name: string, opts: IReadOptions): Promise<string> {
+export async function readParsable(name: string | number, opts: IReadOptions): Promise<string> {
   try {
     return await readFile(name, opts);
   } catch (ex) {
