@@ -11,7 +11,7 @@ function runFalsy(targetVal: any, targetPath?: any) {
     {
       given: null,
       original: null,
-    },
+    } as any,
   );
 }
 
@@ -27,7 +27,7 @@ describe('falsy', () => {
   test('returns error message if target value is not falsy', () => {
     expect(runFalsy(true)).toEqual([
       {
-        message: 'property is not falsy',
+        message: '{{property|gravis|append-property}}is not falsy',
       },
     ]);
   });
