@@ -1,12 +1,10 @@
 import { schemaPath } from '../schema-path';
 
 function runSchemaPath(target: any, field: string, schemaPathStr: string) {
-  return schemaPath(
-    target,
-    { field, schemaPath: schemaPathStr },
-    { given: [], target: [] },
-    { given: null, original: target },
-  );
+  return schemaPath(target, { field, schemaPath: schemaPathStr }, { given: [], target: [] }, {
+    given: null,
+    original: target,
+  } as any);
 }
 
 describe('schema', () => {
@@ -62,7 +60,7 @@ describe('schema', () => {
     expect(runSchemaPath(target, fieldToCheck, path)).toEqual([
       {
         path: ['example'],
-        message: `"example" property should have required property 'url'`,
+        message: 'object should have required property `url`',
       },
     ]);
   });
@@ -78,7 +76,7 @@ describe('schema', () => {
 
     expect(runSchemaPath(target, fieldToCheck, path)).toEqual([
       {
-        message: '"example" property format should match format "url"',
+        message: '{{property|gravis|append-property|optional-typeof}}format should match format `url`',
         path: ['example'],
       },
     ]);
