@@ -21,9 +21,6 @@ export const schemaPath: IFunction<ISchemaPathOptions> = (targetVal, opts, paths
   const relevantObject = opts.field ? object[opts.field] : object;
   if (!relevantObject) return [];
   const { target, given } = paths;
-  const lastItem = target
-    ? target.length > 0 && target[target.length - 1]
-    : given.length > 0 && given[given.length - 1];
 
   // The subsection of the targetValue which contains the schema for us to validate the good bit against
   let schemaObject;
@@ -40,15 +37,5 @@ export const schemaPath: IFunction<ISchemaPathOptions> = (targetVal, opts, paths
     }
   }
 
-  const errors = schema(relevantObject, { schema: schemaObject }, paths, otherValues);
-  return (
-    errors &&
-    errors.map(error => {
-      const propertyPath = [lastItem, opts.field].filter(Boolean).join('.');
-      return {
-        ...error,
-        message: `${propertyPath ? `"${propertyPath}"` : ''} property ${error.message}`,
-      };
-    })
-  );
+  return schema(relevantObject, { schema: schemaObject }, paths, otherValues);
 };
