@@ -1,26 +1,14 @@
+import { createResolveHttp, resolveFile } from '@stoplight/json-ref-readers';
 import { Resolver } from '@stoplight/json-ref-resolver';
-import * as fs from 'fs';
+import { DEFAULT_REQUEST_OPTIONS } from '../request';
 
-import { httpReader } from './http';
+const resolveHttp = createResolveHttp(DEFAULT_REQUEST_OPTIONS);
 
 // resolves files, http and https $refs, and internal $refs
 export const httpAndFileResolver = new Resolver({
   resolvers: {
-    https: httpReader,
-    http: httpReader,
-    file: {
-      resolve(ref: any) {
-        return new Promise((resolve, reject) => {
-          const path = ref.path();
-          fs.readFile(path, 'utf8', (err, data) => {
-            if (err) {
-              reject(err);
-            } else {
-              resolve(data);
-            }
-          });
-        });
-      },
-    },
+    https: { resolve: resolveHttp },
+    http: { resolve: resolveHttp },
+    file: { resolve: resolveFile },
   },
 });
