@@ -2,7 +2,7 @@ import { readRuleset } from '../reader';
 
 describe('Rulesets reader', () => {
   it('should resolve oas2-schema', async () => {
-    const { rules } = await readRuleset('spectral:oas2');
+    const { rules } = await readRuleset('spectral:oas');
     expect(rules['oas2-schema']).not.toHaveProperty('then.functionOptions.schema.$ref');
     expect(rules['oas2-schema']).toHaveProperty(
       'then.functionOptions.schema',
@@ -15,7 +15,7 @@ describe('Rulesets reader', () => {
   });
 
   it('should resolve oas3-schema', async () => {
-    const { rules } = await readRuleset('spectral:oas3');
+    const { rules } = await readRuleset('spectral:oas');
     expect(rules['oas3-schema']).not.toHaveProperty('then.functionOptions.schema.$ref');
     expect(rules['oas3-schema']).toHaveProperty(
       'then.functionOptions.schema',
