@@ -10,9 +10,17 @@ describe('Ruleset Validation', () => {
     expect(assertValidRuleset.bind(null, 'true')).toThrow('Provided ruleset is not an object');
   });
 
-  it('given object with no rules property should throw', () => {
-    expect(assertValidRuleset.bind(null, {})).toThrow('Ruleset must have rules property');
-    expect(assertValidRuleset.bind(null, { rule: {} })).toThrow('Ruleset must have rules property');
+  it('given object with no rules and no extends properties should throw', () => {
+    expect(assertValidRuleset.bind(null, {})).toThrow('Ruleset must have rules or extends property');
+    expect(assertValidRuleset.bind(null, { rule: {} })).toThrow('Ruleset must have rules or extends property');
+  });
+
+  it('given object with extends property only should emit no errors', () => {
+    expect(assertValidRuleset.bind(null, { extends: [] })).not.toThrow();
+  });
+
+  it('given object with rules property only should emit no errors', () => {
+    expect(assertValidRuleset.bind(null, { rules: {} })).not.toThrow();
   });
 
   it('given invalid ruleset should throw', () => {
@@ -181,7 +189,10 @@ describe('Ruleset Validation', () => {
   it('recognizes valid array of functions with object only', () => {
     expect(
       assertValidRuleset.bind(null, {
-        functions: [['foo', {}], ['baz', {}]],
+        functions: [
+          ['foo', {}],
+          ['baz', {}],
+        ],
         rules: {},
       }),
     ).not.toThrow();
@@ -237,7 +248,7 @@ describe('Function Validation', () => {
   it('throws if options supplied to fn does not meet schema', () => {
     const schema: JSONSchema7 = { type: 'string' };
     const wrapped = decorateIFunctionWithSchemaValidation(jest.fn(), schema);
-    expect(() => wrapped({}, 2, { given: [] }, { original: [], given: [] })).toThrow(ValidationError);
+    expect(() => wrapped({}, 2, { given: [] }, { original: [], given: [] } as any)).toThrow(ValidationError);
   });
 
   it('does not call supplied fn if options do not meet schema', () => {
@@ -245,20 +256,20 @@ describe('Function Validation', () => {
     const fn = jest.fn();
     const wrapped = decorateIFunctionWithSchemaValidation(fn, schema);
     try {
-      wrapped({}, 2, { given: [] }, { original: [], given: [] });
+      wrapped({}, 2, { given: [] }, { original: [], given: [] } as any);
     } catch {
       // will throw
     }
 
     expect(fn).not.toHaveBeenCalled();
-    expect(() => wrapped({}, {}, { given: [] }, { original: [], given: [] })).toThrow(ValidationError);
+    expect(() => wrapped({}, {}, { given: [] }, { original: [], given: [] } as any)).toThrow(ValidationError);
   });
 
   it('calls supplied fn and passes all other arguments if options do match schema', () => {
     const schema: JSONSchema7 = { type: 'string' };
     const fn = jest.fn();
     const wrapped = decorateIFunctionWithSchemaValidation(fn, schema);
-    wrapped({}, '2', { given: [] }, { original: [], given: [] });
+    wrapped({}, '2', { given: [] }, { original: [], given: [] } as any);
 
     expect(fn).toHaveBeenCalledWith({}, '2', { given: [] }, { original: [], given: [] });
   });
