@@ -1,9 +1,8 @@
 import * as path from '@stoplight/path';
 import * as fs from 'fs';
-import { filesMap } from './map';
+import { RESOLVE_ALIASES, STATIC_ASSETS } from '../assets';
 
 const SPECTRAL_SRC_ROOT = path.join(__dirname, '..');
-
 // DON'T RENAME THIS FUNCTION, you can move it within this file, but it must be kept as top-level declaration
 // parameter can be renamed, but don't this if you don't need to
 function resolveSpectralVersion(pkg: string) {
@@ -32,6 +31,12 @@ async function resolveFromFS(from: string, to: string) {
   }
 
   targetPath = path.resolve(from, to);
+
+  // if found in static assets, it's fine, as readParsable will handle it just fine
+  if (targetPath in STATIC_ASSETS) {
+    return targetPath;
+  }
+
   // if it's not a built-in ruleset, try to resolve the file according to the provided path
   if (await exists(targetPath)) {
     return targetPath;
@@ -41,9 +46,14 @@ async function resolveFromFS(from: string, to: string) {
 }
 
 export async function findFile(from: string, to: string) {
-  const mapped = filesMap.get(to);
+  const mapped = RESOLVE_ALIASES[to];
+
   if (mapped !== void 0) {
     to = mapped;
+  }
+
+  if (to in STATIC_ASSETS) {
+    return to;
   }
 
   if (path.isAbsolute(to)) {
