@@ -1,5 +1,6 @@
 import { DiagnosticSeverity } from '@stoplight/types';
 import { IRule } from '../../../types';
+import { FileRuleCollection } from '../../../types/ruleset';
 import { mergeRules } from '../rules';
 
 describe('Ruleset rules merging', () => {
@@ -146,18 +147,6 @@ describe('Ruleset rules merging', () => {
     expect(rules).toHaveProperty('test.severity', -1);
   });
 
-  it('supports array-ish syntax', () => {
-    const rules = {
-      test: JSON.parse(JSON.stringify(baseRule)),
-    };
-
-    mergeRules(rules, {
-      test: ['off'],
-    });
-
-    expect(rules).toHaveProperty('test.severity', -1);
-  });
-
   it('does not set functionOptions if rule does not implement it', () => {
     const rules = {
       test: JSON.parse(JSON.stringify(baseRule)),
@@ -203,6 +192,9 @@ describe('Ruleset rules merging', () => {
   it('picks up recommended rules', () => {
     const rules = {};
 
+    const test3 = JSON.parse(JSON.stringify(baseRule));
+    delete test3.recommended;
+
     mergeRules(
       rules,
       {
@@ -211,12 +203,49 @@ describe('Ruleset rules merging', () => {
           ...JSON.parse(JSON.stringify(baseRule)),
           recommended: false,
         },
+        test3,
       },
       'recommended',
     );
 
     expect(rules).toHaveProperty('test.severity', DiagnosticSeverity.Warning);
     expect(rules).toHaveProperty('test2.severity', -1);
+    expect(rules).toHaveProperty('test3.severity', DiagnosticSeverity.Warning);
+  });
+
+  it('rules with no severity and no recommended set are treated as warnings', () => {
+    const rules = {};
+
+    const rule = JSON.parse(JSON.stringify(baseRule));
+    delete rule.recommended;
+    delete rule.severity;
+
+    mergeRules(
+      rules,
+      {
+        rule,
+      },
+      'recommended',
+    );
+
+    expect(rules).toHaveProperty('rule.severity', DiagnosticSeverity.Warning);
+  });
+
+  it('rules with recommended set to false are disabled', () => {
+    const rules = {};
+
+    mergeRules(
+      rules,
+      {
+        rule: {
+          ...JSON.parse(JSON.stringify(baseRule)),
+          recommended: false,
+        },
+      },
+      'recommended',
+    );
+
+    expect(rules).toHaveProperty('rule.severity', -1);
   });
 
   it('sets warning as default severity level if a rule has no severity specified', () => {
@@ -282,6 +311,121 @@ describe('Ruleset rules merging', () => {
     );
 
     expect(rules).toHaveProperty('test.severity', DiagnosticSeverity.Hint);
+  });
+
+  describe('inheriting rules with no explicit severity levels', () => {
+    let rules: FileRuleCollection;
+
+    beforeEach(() => {
+      rules = {
+        rule: {
+          given: '',
+          then: { function: '' },
+          recommended: true,
+        },
+        'rule-with-no-recommended': {
+          given: '',
+          then: { function: '' },
+        },
+        'optional-rule': {
+          given: '',
+          then: { function: '' },
+          recommended: false,
+        },
+      };
+    });
+
+    it('sets warning as a default', () => {
+      const newRules = mergeRules({}, rules);
+
+      const custom = {
+        rule: true,
+        'rule-with-no-recommended': true,
+        'optional-rule': true,
+      };
+
+      expect(mergeRules(newRules, custom)).toEqual({
+        rule: expect.objectContaining({
+          recommended: true,
+          severity: DiagnosticSeverity.Warning,
+        }),
+        'rule-with-no-recommended': expect.objectContaining({
+          severity: DiagnosticSeverity.Warning,
+        }),
+        'optional-rule': expect.objectContaining({
+          recommended: false,
+          severity: DiagnosticSeverity.Warning,
+        }),
+      });
+    });
+
+    it('sets warning as a default even when all rules were disabled', () => {
+      const newRules = mergeRules({}, rules, 'off');
+
+      const custom = {
+        rule: true,
+        'rule-with-no-recommended': true,
+        'optional-rule': true,
+      };
+
+      expect(mergeRules(newRules, custom)).toEqual({
+        rule: expect.objectContaining({
+          recommended: true,
+          severity: DiagnosticSeverity.Warning,
+        }),
+        'rule-with-no-recommended': expect.objectContaining({
+          severity: DiagnosticSeverity.Warning,
+        }),
+        'optional-rule': expect.objectContaining({
+          recommended: false,
+          severity: DiagnosticSeverity.Warning,
+        }),
+      });
+    });
+
+    it('respects ruleset severity', () => {
+      expect(mergeRules({}, rules, 'all')).toEqual({
+        rule: expect.objectContaining({
+          recommended: true,
+          severity: DiagnosticSeverity.Warning,
+        }),
+        'rule-with-no-recommended': expect.objectContaining({
+          severity: DiagnosticSeverity.Warning,
+        }),
+        'optional-rule': expect.objectContaining({
+          recommended: false,
+          severity: DiagnosticSeverity.Warning,
+        }),
+      });
+
+      expect(mergeRules({}, rules, 'recommended')).toEqual({
+        rule: expect.objectContaining({
+          recommended: true,
+          severity: DiagnosticSeverity.Warning,
+        }),
+        'rule-with-no-recommended': expect.objectContaining({
+          severity: DiagnosticSeverity.Warning,
+        }),
+        'optional-rule': expect.objectContaining({
+          recommended: false,
+          severity: -1,
+        }),
+      });
+
+      expect(mergeRules({}, rules, 'off')).toEqual({
+        rule: expect.objectContaining({
+          recommended: true,
+          severity: -1,
+        }),
+        'rule-with-no-recommended': expect.objectContaining({
+          severity: -1,
+        }),
+        'optional-rule': expect.objectContaining({
+          recommended: false,
+          severity: -1,
+        }),
+      });
+    });
   });
 
   it('given invalid rule value should throw', () => {
