@@ -1,8 +1,7 @@
 import { Dictionary } from '@stoplight/types';
 import { DiagnosticSeverity } from '@stoplight/types';
-import { JSONSchema4, JSONSchema6, JSONSchema7 } from 'json-schema';
 import { HumanReadableDiagnosticSeverity, Rule } from './rule';
-import { RuleCollection } from './spectral';
+import { JSONSchema, RuleCollection } from './spectral';
 
 export type FileRuleSeverity = DiagnosticSeverity | HumanReadableDiagnosticSeverity | boolean;
 export type FileRulesetSeverity = 'off' | 'recommended' | 'all';
@@ -14,7 +13,7 @@ export type FileRuleCollection = Dictionary<FileRule, string>;
 export interface IRulesetFunctionDefinition {
   code?: string;
   ref?: string;
-  schema: JSONSchema7 | null;
+  schema: JSONSchema | null;
   name: string;
 }
 
@@ -28,7 +27,7 @@ export interface IRuleset {
 export interface IRulesetFile {
   extends?: Array<string | [string, FileRulesetSeverity]>;
   formats?: string[];
-  rules: FileRuleCollection;
+  rules?: FileRuleCollection;
   functionsDir?: string;
-  functions?: Array<string | [string, JSONSchema4 | JSONSchema6 | JSONSchema7]>;
+  functions?: Array<string | [string, JSONSchema]>;
 }
